Handle dataset differently to support word2vec training
Ideally, the word2vec embeddings learned from the dataset
should only be trained on the train dataset, not on
the entire dataset.

However, because the underlying dataset is not shuffled,
doing a 80/20 split is not likely to have a good balance
of human/gpt data. Therefore the following patch does the following:

1) Using the data helpers, we can read the entire csv file,
perform a shuffle, and create a train and split dataset.
2) Word2Vec is now being trained on the entire dataset
resulting in only one embedding for both human
and gpt data. This is reflected in the training
as well as the embedders.
3) Moved the seed constants and split information
to a constants file.