Adding a dataset that converts all embeddings at once
Given the size of the dataset, the first implementation
of the Pytorch dataset was calculating embeddings on the fly.

However, we know that we have roughly 1,500,000 text pieces.
Using vectors of length 100, where each vector is a float,
this means each sample can be represented with 400 bytes
in vector form. This will yield 4.8GBs of data on memory.

Currently, google collab seems to be able to handle such
amount of data. I still believe calculating the embeddings
once is the answer, but if the memory gets too expensive
we might process text samples in batch, store them
on disk, and load them lazily on the dataset on demand.