Modified HumanChatbot datasets for better training
The following patch enforces a policy on the datasets used for training.

All labels should be 0 based, and ascending. As a result, if a dataset
contains 0 for human and 2 for gpt, the dataset will instead have 0, 1

Why do we do this? Simple: classification on two classes can be easily
done using sigmoid instead of softmax.

Modified adaboost to interact directly with the dataset class instead
of dealing with underlying df to better enforce policies.