Better generalization for multiclass problems
The following allows any of the neural networks to work
as a multiclass classifier by using softmax directly
instead of sigmoid.

During validation, we make sure to normalize using softmax
again. This is not a problem as softmax is proportional,
the only thing we might be doing is extra work. However,
it is better to apply softmax for an accurate prediction.