Compute images on the fly when requested by dataloader.
The image dataset is huge. For a resized of 100x100, this
dataset is 15GB. For an unresized dataset, it is close
to 50GB. This ammount of information has been proved to be
challenging to upload to google drive, specially
because it is many small files, instead of one big file.

Therefore, pivot to generate the images on the fly
by using a wrapper-style dataset class instead
of a imagefolder.