Modified CNN to handle more complex structures
The following patch modifies the CNN to handle
more complex data. After testing a 1 convolutional
layer and seeing a poor performance, this patch
makes the CNN use 2 convolutional blocks
with sigmoid instead of ReLU. The reason for the sigmoid
is that the images are normalized, so values
are between [0, 1]. Sigmoid should be able to better
handle the gradient for these values.

Also, this patch also hardens the validate_after_epoch
method given that we have to call model.eval()
for networks that use batchnorm.